Furiganaをコンポーネントに切り出し
